requirements.txt
pyautogui
tkinter
Pillow